Make adhoc scripts use typer
